Use project name instead of id as blob root key
This allows us to:
1. have user friendly blob URLs, e.g. firmware zips etc
2. have stable blob keys for public projects. simplifies filling
   public projects blob data at startup since we don't have to figure
   out what's the internal project ID.

Downside: project renames will need to handle blob rename. We'll split our
blobstore in a low-level KV DB impl and a more filesystem like API on
top of it, which will have the notion of directories and will be able to
easily rename directories.

For now the blob keys are `user-foo` instead of `user/foo`in order to finish
this migration in a reasonable time. Fuck the yaks. A FS-like API will
make it trivial for us to just use `user/foo` instead and group all the
user data under a user "home directory".

PUBLISHED_FROM=a57183c5c40fbbbdd6dd051f7c064bf38cde4561