Implement moncruise, step 1
Moncruise is a service that dynamically sizes a group of VMs.
It's primary goal is to dynamically size the CI workers so we reduce the
spend on VMs since they are idle most of the day.

This first step defines the basic structure. There is no actual changes
of GCE VM state but we have:

1. reading and keeping up to date with modified config files
2. reading current state of the group of VMs via the GCE API
3. compute a hopefully reasonable size based on the current time and our github
   punchcard data.
4. show real current group info via a miot cloud call.

TODO:

- start/stop VMs according to the desired size
- fetch the desired size from a prometheus metric
- allow manual override of size via a miot cloud call
- ansible conf in prod that pushes this service to the cloud-ci host
  and defines an ourci.yaml file
- automatically create (clone) missing VMs; e.g. if the configured
  maxSize is N but there are fewer VMs labeled as per configuration,
  create the missing ones. Same goes for removal of extra ones.
- when choosing which VM to stop, try to be smart and prefer killing
  those that are idle.

PUBLISHED_FROM=216b4392e7df9ab6fa01ee20317171d084489640