Use filesystem service and report progress
Just for some context about the FS service:
The filesystem service is much more efficient and resilient to large
uploads. The uploaded stuff can be hand-inspected and files can be
written in chunks.

This change makes `miot` use the FS as backing store for the "worspace"
and makes the `Workspace.Download` use the filesystem as well.

Deployer service and fwbuild service needed a couple of tweaks because
they weren't completely transparent (e.g. interpreted the first path
component as project or user ID, which is now impossible due to FS
file naming rules).

The old legacy code is kept until we release this new `miot`; we can
ditch the old code safely later.

I tested the upload itself up to 15 Mb and it seems to work just fine;
however I limited the size of the uploaded sources to 1M.

The limitation can be overridden with a cmdline flag. It might look a little bit
paranoid, but if `miot cloud build` ships up a lot of data then
something is wrong (e.g. we should be smarter about which files to
ignore or support ignore-files, like mbed does).

Some people have slow networks; with a 15 mb file, I go from 16 seconds
upload with my 4g phone, to 16 minutes upload with ADSL; our evauators
would be upset if they have to wait but they might not realize that it's
not the intended behaviour.

--

This commit also sets a bunch of services on our prod machine to
`latest`; we don't have a staging infrastructure yet; so the production
will be staging for fwbuild for a while.

PUBLISHED_FROM=4d22c4eb241ab83091c18b3b217c2aed303f1df9